--- src/main/java/com/sk/cloudmvc/controller/LoginController.java
@@ -1,47 +1,0 @@
-package com.sk.cloudmvc.controller;
-
-import com.sk.cloudmvc.model.User;
-import com.sk.cloudmvc.service.LoginService;
-import com.sk.cloudmvc.until.CommonResult;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.Map;
-
-/**
- * @author qiaochunxiang
- * @date 2020/3/24 14:55
- */
-@Api(value = "/user")
-@RestController
-public class LoginController {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(LoginController.class);
-
-    @Autowired
-    private LoginService loginService;
-
-    @PostMapping("/login")
-    @ApiOperation(value = "用户登录", notes = "用户登录")
-    public CommonResult login(@RequestBody Map<String, Object> jsondata) {
-        CommonResult result = new CommonResult();
-        try {
-            User user = loginService.login(jsondata);
-            if (user != null){
-                result.setData(user);
-            } else{
-                result.setData(false);
-            }
-        } catch (Exception e) {
-            result.setState(500);
-            result.setData(e);
-            LOGGER.error(e.toString(), e);
-        }
-        return result;
-    }
-
-}
